backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Mithun Sri <mithunsrina@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
pre-commit = "^4.0.1"
ruff = "^0.8.3"
django-cors-headers = "^4.6.0"
biopython = "^1.84"
python-decouple = "^3.8"
pandas = "^2.2.3"
pysam = "^0.23.0"
scipy = "^1.15.3"
statsmodels = "^0.14.4"
hmmlearn = "^0.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
